% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Implementation of Knapsack Algorithm}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame, contains variables weight(w) and value(v)}

\item{W}{weight threshold for the knapsack}
}
\value{
Total and Maximum value within the threshold of W
}
\description{
Dynamic Implementation of Knapsack Algorithm
}
\examples{
knapsack_dynamic(x = knapsack_objects[1:100,], W = 3500)
knapsack_dynamic(x = knapsack_objects[1:500,], W = 2000)
}
\seealso{
\code{\link{brute_force_knapsack}}
\code{\link{greedy_knapsack}}
}
