% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Example data for the knapsack problem}
\format{
## `knapsack_objects`
A data frame with 2000 rows and 2 columns
\describe{
\item{w}{weight}
\item{v}{value}
...}
}
\source{
\url{https://www.ida.liu.se/~732A94/}
}
\usage{
knapsack_objects
}
\description{
Example data for the knapsack problem
}
\keyword{datasets}
