% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsacked-package.R
\docType{package}
\name{knapsacked-package}
\alias{knapsacked}
\alias{knapsacked-package}
\title{knapsacked: Different ways of solving the knapsack problem}
\description{
This package contains three different algorithms that all calculate the knapsack problem in different ways. The algorithms included is the brute force method with an option to parallelize the computations, dynamic programming and greedy heuristic. They all differ in computational complexity where the brute force is the sloest and greedy heurisitc is the quickest. However, brute force is the only algorithm that guarantees an optimal solution, while greedy heuristic only runs until a certain threshold of the optimal solution is attained.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wountoto/Lab6_R https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}

}
\author{
\strong{Maintainer}: Johannes Hedström \email{johed883@student.liu.se}

Authors:
\itemize{
  \item Mikael Montén \email{mikmo937@student.liu.se}
  \item Siddhesh Sreedar \email{sidsr770@student.liu.se}
}

}
\keyword{internal}
