% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_brute_force.R
\name{slow_brute_force}
\alias{slow_brute_force}
\title{Basic and slow implementation of Brute-Force Algorithm}
\usage{
slow_brute_force(x, W)
}
\arguments{
\item{x}{data.frame, contains variables weight(w) and value(v)}

\item{W}{weight threshold for the knapsack}
}
\value{
Total and Maximum value within the threshold of W
}
\description{
The function returns maximum value within all the possible solutions
}
\examples{

slow_brute_force(x = knapsack_objects[1:8,], W = 3500)
slow_brute_force(x = knapsack_objects[1:12,], W = 3500)
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack problem}
}
\seealso{
\code{\link{brute_force_knapsack}}
}
