% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search with parallelization}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data frame consisting of columns w = weights and v = value for individual objects}

\item{W}{The capacity of the knapsack}

\item{parallel}{Whether parallelization should be used or not. Defaults to FALSE.}
}
\value{
A list containing the maximum value that could be fit in the knapsack and row wise elements in the data that was added together
}
\description{
Brute force search with parallelization
}
\examples{
# This is an example of how to use the function
brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)

}
