% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search with parallelization}
\source{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A data frame consisting of columns w = weights and v = value for individual objects}

\item{W}{The capacity of the knapsack}

\item{parallel}{Whether parallelization should be used or not. Defaults to FALSE.}
}
\value{
A list containing the maximum value that could be fit in the knapsack and row wise elements in the data that was added together
}
\description{
Brute force search with parallelization
}
