% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic search}
\source{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data frame consisting of columns w = weights and v = value for individual objects}

\item{W}{The capacity of the knapsack
#'}
}
\value{
A list containing the maximum value that could be fit in the knapsack and row wise elements in the data that was added together
}
\description{
Greedy heuristic search
}
