% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming search}
\source{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#0.2F1_knapsack_problem}
}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data frame consisting of columns w = weights and v = value for individual objects}

\item{W}{The capacity of the knapsack}
}
\value{
A list containing the maximum value that could be fit in the knapsack and row wise elements in the data that was added together
}
\description{
Dynamic programming search
}
\examples{
# This is an example of how to use the function
knapsack_dynamic(x = knapsack_objects[1:8,], W = 3500)

}
